Fixed some gcc warnings.
Former-commit-id: 47c76baa77acd04b1e71bc7f72cc88abe21e666d
Former-commit-id: b4b548ba4b1899ed5f14da3f524d473ae7089f9e

+++ src/diagnostics/fileUtils.cpp
@@ -9,7 +9,7 @@ void fileUtils::getDirectoryFromPath(const char* fullPath, char* directory)
     auto pathSize = strlen(fullPath);
     auto pathSeparator = L'\\';
 
-    for (int i = 0; i < pathSize; i++)
+    for (size_t i = 0; i < pathSize; i++)
     {
         if (fullPath[i] != pathSeparator)
         {
